\name{bim-internal}
\alias{bim-internal}
\alias{.bmapqtl.options}
\alias{bmapqtl.options.init}
\alias{bim.cex}
\alias{bim.cross}
\alias{bim.legacy}
\alias{bim.match}
\alias{bim.prior}
\alias{bim.smooth}
\alias{plot.bim.pattern}
\title{Internal bim functions.}
\description{  These are generally not to be called by the user.
}
\usage{
.bmapqtl.options
bmapqtl.options.init()
bim.cex( bim )
bim.cross( bim )
bim.legacy( bim )
bim.match( bim, pattern )
bim.prior( bim, range )
bim.smooth( x, y )
plot.bim.pattern( x, bars, labels, barlabels, threshold, units,
  rescale, \dots )
}
\author{ Brian S. Yandell, yandell@stat.wisc.edu }
\keyword{internal}
