% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_best_model.R
\name{choose_best_model}
\alias{choose_best_model}
\title{Choose Predictors}
\usage{
choose_best_model(data, response_var, predictor_vars)
}
\arguments{
\item{data}{A data frame containing the variables.}

\item{response_var}{Variable to model.}

\item{predictor_vars}{Predictor variables.}
}
\description{
This function returns the full model with all predictors, as well as a reduced model using only the best predictors.
}
\seealso{
\href{../doc/Choose_best_model.html}{Choose Best Model Vignette}
}
\keyword{predictors}
