% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats_function.R
\name{descriptive_data}
\alias{descriptive_data}
\title{Provides Descriptive Statistics}
\usage{
descriptive_data(
  data,
  group_variables,
  cell_variables,
  stats = c("n", "median", "q1", "q3", "mean", "sd", "sem")
)
}
\arguments{
\item{data}{A data frame or tibble containing a dataset.}

\item{group_variables}{A character vector of your column names that is written as "text". This will tell the function how to group the data (e.g., group_variables(c("Sex", "Genotype"))).}

\item{cell_variables}{A character vector of column names that is written as "text". This should have the numerical data that you want to summarize (e.g., cell_variables(c("T_Cells", "B_Cells", "Microglia"))).}

\item{stats}{A character vector of which descriptive statistics to calculate. Options include: "n", "median", "q1", "q3", "mean", "sd", "sem" and will defaults to all of them if no stats are chosen.}
}
\value{
A data frame where each row represents a group and each column contains the summary statistics that were calculated.
Each column of descriptive statistics will contain the name of the numerical variable and have the name of the statistic (e.g., T_Cells_mean ).
}
\description{
This function can be used to calculate different types of descriptive statistics (e.g., number of subjects (n), median, interquartile range (IQR), mean, standard deviation (SD), and the standard error of the mean (SEM)).
These descriptive statistics are calculated for any group and for any numerical variables of your choosing.
}
\examples{
descriptive_data(
  data = data_frame,
  group_variables = c("Sex", "Genotype"),
  cell_variables = c("T_Cells", "B_Cells", "Microglia"),
  stats = c("n", "mean", "sd")
)
}
\keyword{descriptive}
\keyword{statistics}
\keyword{summary}
