\name{fisch.bim}
\alias{fisch.bim}
\non_function{}
\title{MCMC samples from posterior for Stephens and Fisch simulation}
\description{
  Samples use exponential prior on number of QTL with prior mean 3.
  See \code{read.bim} for details.
}
\usage{data(fisch.bim)}
\references{ \url{www.stat.wisc.edu/~yandell/qtl/software/Bmapqtl} }
\author{ Brian S. Yandell, yandell@stat.wisc.edu }
\seealso{\code{\link{read.bim}},\code{\link{plot.bim}}}

\examples{
data(fisch.bim)
plot( fisch.bim )
}
\keyword{datasets}
