% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot_function.R
\name{graph_boxplots}
\alias{graph_boxplots}
\title{Creates Boxplot}
\usage{
graph_boxplots(
  data,
  x_variable,
  y_variables,
  fill_variable = NULL,
  facet_variable = NULL,
  add_points = TRUE,
  coef = 1.5
)
}
\arguments{
\item{data}{A data frame or tibble containing a dataset that will be used to make plots.}

\item{x_variable}{The name of the categorical variable to use on the x-axis, written as a string or as "text" (e.g., x_variable("Age")).}

\item{y_variables}{The names of the numerical variables to use on the y-axis, written as character vector or as "text" (e.g., y_variable(c("T_Cells", "B_Cells", "Microglia""))).
A separate plot will be created for each y-variable.}

\item{fill_variable}{(Optional) A variable to use for fill color and point color, written as a string or as "text." If NULL, then no fill will be used.}

\item{facet_variable}{(Optional) A variable to use for faceting the plots, written as a string or as "text." If NULL, then no fill will be used.}

\item{add_points}{(TRUE or FALSE). If TRUE, individual data points will be added to the boxplots. The default is TRUE. If FALSE, then no facet wrap will occur.}

\item{coef}{A numerical value that controls the length of the whiskers in the boxplot. This will be multiplied by the IQR to determine whisker size.
Use (coef = 1.5) for Tukey’s standard whiskers.
Use (coef = Inf) to extend whiskers to the min/max values.
Use (coef = 0) to remove whiskers.}
}
\value{
One or more boxplots will be made for each of y-variables that were given.
}
\description{
This function can be used to create a multiple boxplots for one or more numerical y-variables and can be grouped by any categorical x-variables.
This function can create boxplots with the option to color fill, include raw data points, and facet by any variable.
}
\examples{
p <- graph_boxplots(
  data = data_frame,
  x_variable = "Genotype",
  y_variables = c("T_Cells", "B_Cells", "Microglia"),
  fill_variable = "Time_of_day",
  facet_variable = "Sex",
  add_points = TRUE
)
}
\seealso{
\href{../doc/Boxplot_function.html}{Vignette: Boxplot_function}
}
\keyword{boxplot}
\keyword{plotting}
\keyword{visualization}
