% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_model_table.R
\name{make_model_table}
\alias{make_model_table}
\title{Function to make tables of summary of the model}
\usage{
make_model_table(model)
}
\arguments{
\item{model}{A statistical model of data.}
}
\description{
Table of Estimate, SE, F, and p for the given model
}
\examples{
make_model_table(brain_CD8_model)
}
\seealso{
\link{vignette("bim")}
}
