\name{vern}
\alias{vern}
\non_function{}
\title{Cross structure for 8-week vernalization data}
\description{
Contains genotypes and phenotypes for 8-week vernalization study.
}
\usage{data(vern)}
\format{
See \code{\link[qtl]{read.cross}} in \code{library(qtl)} for format.
}
\seealso{\code{\link[qtl]{read.cross}},\code{\link{plot.bim}}}
\source{
Thomas C. Osborn (tcosborn@facstaff.wisc.edu), Department of Agronomy,
UW-Madison. 
}
\references{
\url{www.stat.wisc.edu/~yandell/qtl/data/osborn/Bnapus}
}
\examples{
data(vern)
}
\keyword{datasets}
