\name{verngeo.bim}
\alias{verngeo.bim}
\non_function{}
\title{MCMC samples from posterior for 8-week vernalization}
\description{
  Samples use geometric prior on number of QTL with prior mean 3.
}
\usage{data(verngeo.bim)}
\references{ \url{www.stat.wisc.edu/~yandell/qtl/software/Bmapqtl} }
\author{ Brian S. Yandell, yandell@stat.wisc.edu }
\seealso{\code{\link{vern}},\code{\link{run.bmapqtl}},\code{\link{plot.bim}}}

\examples{
data(verngeo.bim)
plot( verngeo.bim )
}
\keyword{datasets}
