% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bim-package.R, R/choose_best_model.R,
%   R/make_model_table.R, R/plot_model_with_ci.R
\docType{package}
\name{bim-package}
\alias{bim-package}
\title{bim: Brain Immune Migration (bim)}
\description{
This package provides tools to analyze rhythmic patterns of immune cell migration, with a focus on T cells in the brain. It aims to identify whether specific immune cell subtypes are predictive of CD8+ T cell dynamics. The package supports analysis of general immune cell rhythms, characterization of their functional profiles, and visualization through customizable plots.

This package provides tools to analyze rhythmic patterns of immune cell migration, with a focus on T cells in the brain. It aims to identify whether specific immune cell subtypes are predictive of CD8+ T cell dynamics. The package supports analysis of general immune cell rhythms, characterization of their functional profiles, and visualization through customizable plots.

This package provides tools to analyze rhythmic patterns of immune cell migration, with a focus on T cells in the brain. It aims to identify whether specific immune cell subtypes are predictive of CD8+ T cell dynamics. The package supports analysis of general immune cell rhythms, characterization of their functional profiles, and visualization through customizable plots.

This package provides tools to analyze rhythmic patterns of immune cell migration, with a focus on T cells in the brain. It aims to identify whether specific immune cell subtypes are predictive of CD8+ T cell dynamics. The package supports analysis of general immune cell rhythms, characterization of their functional profiles, and visualization through customizable plots.
}
\author{
\strong{Maintainer}: Joyradyn James-Rollins \email{jjamesrollins237@utexas.edu}

Authors:
\itemize{
  \item Andrew Hynes \email{drew.hynes@utexas.edu}
  \item Stephen Collins \email{sacollins@utexas.edu}
}

}
\keyword{internal}
