% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_best_model.R
\name{choose_best_model}
\alias{choose_best_model}
\title{Function to choose predictors in model}
\usage{
choose_best_model(data, response_var, predictor_vars)
}
\arguments{
\item{data}{A data frame containing the variables.}

\item{response_var}{Variable to model.}

\item{predictor_vars}{Predictor variables.}
}
\description{
Gives original model of data with all predictors, model of data with only best predictors
}
\examples{
choose_best_model <- function(data, response_var, predictor_vars)
}
\seealso{
\link{vignette("bim")}
}
