\name{fisch}
\alias{fisch}
\non_function{}
\title{Cross structure for 8 QTL Stephens and Fisch simulated data}
\description{
Contains genotypes and phenotypes for data simulated using model in
Stephens and Fisch (1998) but with 90 percent heritability.
}
\usage{data(fisch)}
\format{
See \code{read.cross} in \code{library(qtl)} for format.
}
\seealso{\code{\link[qtl]{read.cross}},\code{\link{plot.bim}}}
\source{
Patrick J. Gaffney (paga@lubrizol.com), Lubrizol Corp.
}
\references{
\url{www.stat.wisc.edu/~yandell/qtl/software/Bmapqtl}
}
\examples{
data(fisch)
}
\keyword{datasets}
