% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_model_table.R
\name{make_model_table}
\alias{make_model_table}
\title{Creates Table Summary}
\usage{
make_model_table(model)
}
\arguments{
\item{model}{A statistical model of data.}
}
\description{
This function generates a table of estimates, standard errors (SE), F-values, and p-values for the specified model.
}
\examples{
make_model_table(brain_CD8_model)
}
\seealso{
\href{../doc/Make_model_table.html}{Make Model Table Vignette}
}
\keyword{statistical}
\keyword{summary}
